+++ packages/nimble-blazor/NimbleBlazor/Components/NimbleTableColumnText.razor.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Components;
 
 namespace NimbleBlazor;
 
-public partial class NimbleTableColumnText : ComponentBase
+public partial class NimbleTableColumnText : NimbleTableColumn
 {
     /// <summary>
     /// Gets or sets the field in the element representing a row of data in a <see cref="NimbleTable{TData}"/>to display